Convert cassowary algorithm to use Rational type
