+++ src/app/models/data/belt.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../rational';
+import { Rational } from '../math/rational';
 
 export interface Belt {
   speed: number;
