+++ src/app/models/data/factory.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../rational';
+import { Rational } from '../math/rational';
 
 export interface Factory {
   speed: number;
