+++ src/app/models/data/item.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../rational';
+import { Rational } from '../math/rational';
 import { Belt, RationalBelt } from './belt';
 import { CategoryId } from './category';
 import { Factory, RationalFactory } from './factory';
