+++ src/app/models/enum/display-rate.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../rational';
+import { Rational } from '../math/rational';
 
 export enum DisplayRate {
   PerSecond = 1,
