+++ src/app/models/enum/research-speed.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../rational';
+import { Rational } from '../math/rational';
 
 export enum ResearchSpeed {
   Speed0 = 0,
