+++ src/app/models/index.ts
@@ -3,8 +3,8 @@ export * from './enum';
 export * from './constants';
 export * from './entities';
 export * from './item-settings';
+export * from './math';
 export * from './node';
 export * from './product';
-export * from './rational';
 export * from './recipe-settings';
 export * from './step';
