+++ src/app/models/math/maptype.ts
